third_party/blink/web_tests/external/wpt/webidl/ecmascript-binding/constructors-expected.txt

This is a testharness.js-based test.
[FAIL] Constructor in child window with bad NewTarget from parent window
  assert_equals: expected object "[object DOMParser]" but got object "[object Object]"
[FAIL] Constructor in parent window with bad NewTarget from child window
  assert_equals: expected object "[object DOMParser]" but got object "[object Object]"
[FAIL] Constructor in parent window with bad NewTarget from parent window that's a bound child window function
  assert_equals: expected object "[object DOMParser]" but got object "[object Object]"
[FAIL] Constructor in child window with bad NewTarget from child window that's a bound parent window function
  assert_equals: expected object "[object DOMParser]" but got object "[object Object]"
[FAIL] Constructor in parent window with bad NewTarget from parent window that's a proxy for a child window function
  assert_equals: expected object "[object DOMParser]" but got object "[object Object]"
[FAIL] Constructor in child window with bad NewTarget from child window that's a proxy for a parent window function
  assert_equals: expected object "[object DOMParser]" but got object "[object Object]"
Harness: the test ran to completion.
